fixed build-time issues
