+++ app/api/generate-link/route.ts
@@ -1,4 +1,4 @@
-import { NextRequest, NextResponse } from "next/server";
+import { NextResponse } from "next/server";
 import { PrismaClient } from "@prisma/client";
 
 const prisma = new PrismaClient();
@@ -29,7 +29,7 @@ export async function POST() {
     console.error("Error generating link:", error);
     return NextResponse.json(
       { success: false, error: "Failed to generate link" },
-      { status: 500 }
+      { status: 500 },
     );
   }
 }
