+++ app/layout.tsx
@@ -12,12 +12,15 @@ export const metadata: Metadata = {
 
 export default function RootLayout({
   children,
+  modal,
 }: Readonly<{
   children: React.ReactNode;
+  modal: React.ReactNode;
 }>) {
   return (
     <html lang="en">
       <body className={`${sans_serif.className}  antialiased`}>
+        {modal}
         <Toaster />
         <nav>
           <Navbar />
